+++ src/meridian/__init__.py
@@ -2,4 +2,4 @@
 Meridian: Heroku for ML Features.
 """
 
-__version__ = "1.4.0"
+__version__ = "1.5.0"